+++ central/compliance/data/repository_test.go
@@ -4,6 +4,8 @@ import (
 	"bytes"
 	"compress/gzip"
 	"encoding/json"
+	"maps"
+	"slices"
 	"testing"
 
 	"github.com/stackrox/rox/generated/internalapi/compliance"
@@ -12,7 +14,6 @@ import (
 	"github.com/stackrox/rox/pkg/protoassert"
 	"github.com/stackrox/rox/pkg/utils"
 	"github.com/stretchr/testify/suite"
-	"golang.org/x/exp/maps"
 )
 
 func TestRepositorySuite(t *testing.T) {
@@ -84,7 +85,7 @@ func (s *RepositoryTestSuite) TestGetNodeResults() {
 
 	nodeResults := getNodeResults(testScrapeResults)
 
-	s.Equal(maps.Keys(nodeResults), []string{testNodeName})
+	s.Equal(slices.Collect(maps.Keys(nodeResults)), []string{testNodeName})
 	actual := nodeResults[testNodeName]
 	protoassert.MapEqual(s.T(), testEvidence, actual)
 }